[MOV] im_livechat: move Timer out of public_livechat.js
Extract the Timer and Timers utils from the public_livechat.js,
and move it to public_livechat_utils.js.

The goal is to split the public_livechat.js file into multiple parts,
to prepare for refactoring.

Task-2212347
